# perf(bytes): improve performance of `equals()` 