--- a/bytes/equals.ts
+++ b/bytes/equals.ts
@@ -27,17 +27,23 @@ function equalsNaive(a: Uint8Array, b: Uint8Array): boolean {
  */
 function equals32Bit(a: Uint8Array, b: Uint8Array): boolean {
   const len = a.length;
-  const compressible = Math.floor(len / 4);
-  const compressedA = new Uint32Array(a, 0, compressible);
-  const compressedB = new Uint32Array(b, 0, compressible);
-  for (let i = compressible * 4; i < len; i++) {
+  const compactOffset = 3 - ((a.byteOffset + 3) % 4);
+  const compactLen = Math.floor((len - compactOffset) / 4);
+  const compactA = new Uint32Array(a.buffer, a.byteOffset + compactOffset, compactLen);
+  const compactB = new Uint32Array(b.buffer, b.byteOffset + compactOffset, compactLen);
+  for (let i = 0; i < compactOffset; i++) {
     if (a[i] !== b[i]) return false;
   }
-  for (let i = 0; i < compressedA.length; i++) {
-    if (compressedA[i] !== compressedB[i]) return false;
+  for (let i = 0; i < compactA.length; i++) {
+    if (compactA[i] !== compactB[i]) return false;
+  }
+  for (let i = compactOffset + compactLen * 4; i < len; i++) {
+    if (a[i] !== b[i]) return false;
   }
   return true;
 }
+
+const THRESHOLD_32_BIT = 160;
 
 /**
  * Check whether byte slices are equal to each other.
@@ -62,5 +68,7 @@ export function equals(a: Uint8Array, b: Uint8Array): boolean {
   if (a.length !== b.length) {
     return false;
   }
-  return a.length < 1000 ? equalsNaive(a, b) : equals32Bit(a, b);
+  return a.length >= THRESHOLD_32_BIT && (a.byteOffset % 4) === (b.byteOffset % 4)
+    ? equals32Bit(a, b)
+    : equalsNaive(a, b);
 }