--- a/bytes/equals.ts
+++ b/bytes/equals.ts
@@ -52,8 +52,10 @@ function equals32Bit(a: Uint8Array, b: Uint8Array): boolean {
 }
 
 /**
- * Threshold for when to use 32-bit comparisons. This was chosen based on
- * benchmark results.
+ * Byte length threshold for when to use 32-bit comparisons, based on
+ * benchmarks.
+ *
+ * @see {@link https://github.com/denoland/deno_std/pull/4635}
  */
 const THRESHOLD_32_BIT = 160;